requirements.txt
django==3.2.7
Pillow==8.3.2
environs[django]==9.3.3
django-debug-toolbar==3.2.2
gunicorn==20.1.0
django-heroku==0.3.1
django_bootstrap5==2.1.2
django-import-export==2.6.0